fix(control-plane): extract provider capabilities off the frozen desktop backend (TASK-672)
